requirements.txt
# Django & REST Framework
django>=4.2,<5.0
djangorestframework>=3.14
django-filter>=23.0
django-cors-headers>=4.0
python-dotenv>=1.0

# Database
psycopg2-binary>=2.9
dj-database-url>=1.0

# Authentication & Tokens
djangorestframework-simplejwt>=5.2
django-allauth>=0.50

# API Documentation
drf-spectacular>=0.26

# Celery & Redis
celery>=5.3
redis>=4.5
django-celery-beat>=2.5
django-celery-results>=2.5

# Utilities
Pillow>=9.0
python-dateutil>=2.8
pytz>=2023.3
requests>=2.31

# Development
django-extensions>=3.2
black>=23.0
flake8>=6.0
pytest>=7.0
pytest-django>=4.5
pytest-cov>=4.0
factory-boy>=3.2
faker>=18.0

# Production
gunicorn>=20.1
whitenoise>=6.4


